backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.4
django-cors-headers==4.9.0
djangorestframework==3.17.1
drf-spectacular==0.29.0
gunicorn==23.0.0
idna==3.11
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
